+++ sql/emaj--2.3.0--2.3.1.sql
@@ -1,5 +1,5 @@
 --
--- E-Maj: migration from 2.3.0 to <devel>
+-- E-Maj: migration from 2.3.0 to 2.3.1
 --
 -- This software is distributed under the GNU General Public License.
 --
@@ -411,10 +411,10 @@ INSERT INTO pg_catalog.pg_description (objoid, classoid, objsubid, description)
        );
 
 -- update the version id in the emaj_param table
-UPDATE emaj.emaj_param SET param_value_text = '<devel>' WHERE param_key = 'emaj_version';
+UPDATE emaj.emaj_param SET param_value_text = '2.3.1' WHERE param_key = 'emaj_version';
 
 -- insert the upgrade record in the operation history
-INSERT INTO emaj.emaj_hist (hist_function, hist_object, hist_wording) VALUES ('EMAJ_INSTALL','E-Maj <devel>', 'Upgrade from 2.3.0 completed');
+INSERT INTO emaj.emaj_hist (hist_function, hist_object, hist_wording) VALUES ('EMAJ_INSTALL','E-Maj 2.3.1', 'Upgrade from 2.3.0 completed');
 
 -- post installation checks
 DO